+++ lib/ooxml_parser/common_parser/common_data/paragraph/paragraph_run/run_properties.rb
@@ -1,7 +1,7 @@
 require_relative 'run_properties/language'
 require_relative 'run_properties/outline'
 require_relative 'run_properties/position'
-require_relative 'run_properties/run_size'
+require_relative 'run_properties/size'
 require_relative 'run_properties/run_spacing'
 require_relative 'run_properties/shade'
 require_relative 'run_properties/strikeout'
@@ -9,7 +9,7 @@ module OoxmlParser
   class RunProperties < OOXMLDocumentObject
     attr_accessor :font_style, :font_color, :space, :dirty, :font_name, :font_size, :baseline, :hyperlink, :caps,
                   :vertical_align, :outline
-    # @return [RunSize] get run size
+    # @return [Size] get run size
     attr_accessor :size
     # @return [RunSpacing] get run spacing
     attr_accessor :spacing
@@ -64,7 +64,7 @@ module OoxmlParser
       character_props_node.xpath('*').each do |properties_element|
         case properties_element.name
         when 'sz'
-          character_properties.size = RunSize.parse(properties_element)
+          character_properties.size = Size.new.parse(properties_element)
         when 'spacing'
           character_properties.spacing = RunSpacing.parse(properties_element)
         when 'color'